Remove 'related' from parent field
This change makes the overral Entity simpler, but it also makes it less
powerfull.
Entities are now referenced with the name given in the database and
nothing else. The consequente is that there is no concept of relation
between the entities and, therefore, we cannot have entities raferencing
the same other entity twice with different purposes.
An entity 'person' cannot reference other persons as 'colleagues' and
reference them as 'relatives' for example.

This situatios if far from ideal, but the usage of the 'related-' prefix
couldn't really help in the manner that it was implemented.